// src/app/shared/services/wmp-api/models/WmpHistoricProcessInstanceDto.ts
//
//   Copyright 2026 Deutsches Zentrum für Luft- und Raumfahrt e.V.
//
//   Licensed under the Apache License, Version 2.0 (the "License");
//   you may not use this file except in compliance with the License.
//   You may obtain a copy of the License at
//
//       http://www.apache.org/licenses/LICENSE-2.0
//
//   Unless required by applicable law or agreed to in writing, software
//   distributed under the License is distributed on an "AS IS" BASIS,
//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
//   See the License for the specific language governing permissions and
//   limitations under the License.
//

import {HistoricProcessInstanceDto} from 'src/app/shared/services/camunda-api';

export interface WmpHistoricProcessInstanceDto
    extends HistoricProcessInstanceDto {
    /**
     * The time of last activity (start/end) for this process instance `yyyy-MM-dd\'T\'HH:mm:ss.SSSZ`.
     */
    latestStartTime?: string | null;

    latestEndTime?: string | null;

    incidentMessage?: string | null;

    incidentJobId?: string | null;
    deploymentId?: string | null;
}